The commit message was empty

+++ src/edu/mit/csail/sdg/alloy4/OurImageUtil.java
@@ -41,13 +41,13 @@ public final class OurImageUtil {
     private static final int BUFSIZE=4096;
 
     /** Writes the image as a PNG file with the given horizontal and vertical dots-per-inch. */
-    public static void writePNG(BufferedImage image, String filename, int dpiX, int dpiY) throws IOException {
+    public static void writePNG(BufferedImage image, String filename, double dpiX, double dpiY) throws IOException {
         ImageIO.write(image, "PNG", new File(filename));
         setDPI(filename, dpiX, dpiY);
     }
 
     /** Modifies the given PNG file to have the given horizontal and vertical dots-per-inch. */
-    private static void setDPI(String filename, int dpiX, int dpiY) throws IOException {
+    private static void setDPI(String filename, double dpiX, double dpiY) throws IOException {
        RandomAccessFile f = null;
        try {
           f = new RandomAccessFile(filename, "rw");
@@ -111,9 +111,9 @@ public final class OurImageUtil {
     }
 
     /** Write the "pHYs" chunk into the PNG file with the given horizontal and vertical dots-per-inch. */
-    private static void writeDPI(RandomAccessFile f, int dpiX, int dpiY) throws IOException {
-        int dpmX = (int) (((double)dpiX)/2.54d*100d); // Translate dots-per-inch into dots-per-meter
-        int dpmY = (int) (((double)dpiY)/2.54d*100d); // Translate dots-per-inch into dots-per-meter
+    private static void writeDPI(RandomAccessFile f, double dpiX, double dpiY) throws IOException {
+        int dpmX = (int) (dpiX/2.54d*100d); // Translate dots-per-inch into dots-per-meter
+        int dpmY = (int) (dpiY/2.54d*100d); // Translate dots-per-inch into dots-per-meter
         int crc = 0xFFFFFFFF, b;
         f.write(0); f.write(0); f.write(0); f.write(9);
         b='p';              crc=table[(crc ^ b) & 0xff] ^ (crc >>> 8); f.write(b);
